# Add DroppedAttributeCount field to Otel/STEF schema