… benchmarks/benchmarks_test.go
… benchmarks/readwrite_test.go
… benchmarks/size_test.go
… benchmarks/testdata/astronomyshop.stefz
… benchmarks/testdata/astronomyshop.tefz
… benchmarks/testdata/hipstershop.stefz
… benchmarks/testdata/hipstershop.tefz
… benchmarks/testdata/hostandcollectormetrics.stefz
… benchmarks/testdata/hostandcollectormetrics.tefz
… benchmarks/testdata/oteldemo-with-histogram.stefz
… benchmarks/testdata/oteldemo-with-histogram.tefz
… go/otel/oteltef.wire.json
… go/otel/oteltef/metricswriter.go
… go/otel/oteltef/resource.go
… go/otel/oteltef/scope.go
… go/otel/oteltef/span.go
… go/otel/oteltef/spans.go
… go/otel/oteltef/spanstatus.go
… go/otel/oteltef/spanswriter.go
… go/pdata/internal/otlptools/convert.go
… go/pdata/internal/otlptools/otlpval2tef.go
… go/pdata/metrics/testtools/normalize.go
… go/pdata/traces/otlp2tef.go
… stefgen/templates/array.go.tmpl
… stefgen/templates/multimap.go.tmpl
… stefgen/templates/oneof.go.tmpl
… stefgen/templates/struct.go.tmpl