Resolves https://github.com/splunk/stef/issues/8

Otel/STEF schema changes:

- Added "DroppedAttributeCount" field in the "Resource", "Scope", "Span", etc structs.

Also updated testdata files (because the schema is now different) and renamed the incorrect ".tefz" to ".stefz" extension.