@dmitryax 2025-01-27
Just to clarify, why does the dropped attributes field need to be embedded in the `Attributes`?