@tigrannajaryan 2025-01-27
@dmitryax BTW, one other somewhat related thing to keep in mind is how STEF handles wire format compatibility.

Unlike ProtoBuf the only way to evolve STEF schema in a way that is backward compatible is by adding new fields at the very end of "struct" or "oneof" types. Adding new fields in the "middle" of a struct or oneof breaks wire compatibility in STEF (but is not a problem in ProtoBuf since the fields are numbered). STEF allows adding such fields at the end of any struct, including nested structs, not just the root struct.

Because of this limitation it is beneficial to design the initial schema in a way that logically related fields are grouped into their own struct, so that if another logically related field needs to be added it can be added to the relevant nested struct, instead of being added as the last struct in a large flat root struct.

I don't think we will want to adding any other fields that are related to Attributes in the future, but if we ever did then having "Attributes" as a separate struct would result in a bit nicer schema. Not a dealbreaker, just a nice to have thing.