@dmitryax 2025-01-28
Thanks for clarifying. That sounds good to me, assuming that there is no performance overhead of introducing the wrapping struct