@tigrannajaryan 2025-01-29
@dmitryax I spent a bit more time thinking about this and I think it is best to keep the Otel/STEF schema as close as possible to OTLP Protobufs. I will switch to the sibling schema for DroppedAttributesCount.